chatbot_requirements.txt
streamlit
google-generativeai==0.1.0
